ml-service/requirements.txt
psycopg2-binary==2.9.9
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0